Reengineer connectivity checking
Instead of finding the entire contiguous block (may be wasteful in some
cases) we now use a function that asks if two given sets are connected
by a contiguous block containing a given symbol. This function can
terminate the search early, as soon as it finds a connection, and not
waste its time finding more of the contiguous block.